Fix build errors: refactor glossary data, fix imports and types
