+++ Web_Frontend/genfit_frontend/src/lib/utils/chatLinkParser.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { glossaryExercises } from '../../pages/knowledge-hub/GlossaryPage';
+import { glossaryExercises } from '../../pages/knowledge-hub/glossaryData';
 
 
 /**
